Fix type conversion errors in WebSocket/HTTP code
Fixed multiple type-related issues:

1. C wrapper type mismatch:
   - Changed lws_msg_len() return type from size_t to int
   - Matches Hemlock extern declaration (i32)

2. Incorrect use of typeof() for value conversion:
   - typeof() returns type name ("i32"), not the value
   - Hemlock auto-converts numbers to strings in concatenation
   - Fixed in websocket.hml:
     * Binary message length display
     * Server error messages
     * WebSocket URL construction
   - Fixed in test_http_requests.hml:
     * HTTP status code printing
     * Response body length display

These fixes resolve "Cannot convert to int" and "Cannot convert type to target
type" runtime errors.

+++ stdlib/c/lws_wrapper.c
@@ -627,8 +627,8 @@ const unsigned char* lws_msg_binary(ws_message_t *msg) {
     return msg ? msg->data : NULL;
 }
 
-size_t lws_msg_len(ws_message_t *msg) {
-    return msg ? msg->len : 0;
+int lws_msg_len(ws_message_t *msg) {
+    return msg ? (int)msg->len : 0;
 }
 
 void lws_msg_free(ws_message_t *msg) {
